+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - New validation [`commonFileWarnings`](docs/validations.md#commonfilewarnings) to list all file lines containing the word `warning` (useful to surface linting/test warnings).
+- New validation [`commonFileContains`](docs/validations.md#commonfilecontains) to make sure the file contents match a regex.
 
 ## [1.2.0] - 2018-10-31